Ramprasad get priceDetails

+++ src/main/java/com/POSG/PosG/Price/PriceRepository.java
@@ -2,6 +2,6 @@ package com.POSG.PosG.Price;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PriceRepository extends JpaRepository<PriceModel,Integer> {
+public interface PriceRepository extends JpaRepository<PriceModel,Long> {
 
 }
